Move the most expensive tests in HttpServerTest to separate test classes
Rewrite test methods to reuse generated crypto material.
Rewrite moved tests to use Junit 5.
